+++ src/util/is_object_literal.ts
@@ -1,0 +1,5 @@
+export function isObjectLiteral(
+    anything: unknown
+): anything is Record<string, unknown> {
+    return Boolean(anything) && anything.constructor === Object;
+}